backend/app/services/ai_service.py
import os
import google.generativeai as genai
from dotenv import load_dotenv

load_dotenv()

class AIService:
    def __init__(self):
        api_key = os.getenv("GEMINI_API_KEY")
        if not api_key:
            print("Warning: GEMINI_API_KEY not set")
            # raise ValueError("GEMINI_API_KEY not set")
            self.model = None
        else:
            genai.configure(api_key=api_key)
            self.model = genai.GenerativeModel('gemini-1.5-pro-latest') # Updating to 1.5 Pro as 3 Pro is likely not public yet or typo in user request (assuming 1.5 Pro or Ultra equivalent)

    async def analyze_image(self, image_path: str):
        if not self.model:
             return {
                "error": "AI Model not configured",
                "latex": "E = mc^2 (Mock)",
                "difficulty": 1,
                "analysis": "Mock analysis due to missing API Key"
            }
        
        # TODO: Implement actual Gemini call
        # image = Image.open(image_path)
        # response = self.model.generate_content(["Analyze this math problem...", image])
        # return parse_response(response.text)
        
        print(f"Mock analyzing image: {image_path}")
        return {
            "latex": "\\int_{0}^{\\infty} x^2 dx",
            "difficulty": 3,
            "analysis": "This is a calculus problem."
        }
