backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
google-generativeai==0.4.0
python-multipart==0.0.9
watchdog==4.0.0
httpx==0.27.0
